requirements.txt
flask
flask-sqlalchemy
boto3
python-dotenv
psycopg2-binary